// src/queries/meal/type.ts
export interface MealResponse {
  data: [
    {
      id: number;
      date: string;
      time: string;
      meal: "";
      calorie: number;
    },
    {
      id: number;
      date: string;
      time: string;
      meal: "";
      calorie: number;
    },
    {
      id: number;
      date: string;
      time: string;
      meal: "";
      calorie: number;
    }
  ];
}

export interface mealOne {
  data: {
    id: number;
    date: string;
    time: string;
    meal: "";
    calorie: number;
  };
}

export interface MealResponse2 {
  
  dataa: [
    {
      id: number;
      date: string;
      time: string;
      meal: "";
      calorie: number;
    },
    {
      id: number;
      date: string;
      time: string;
      meal: "";
      calorie: number;
    },
    {
      id: number;
      date: string;
      time: string;
      meal: "";
      calorie: number;
    }
  
  ];
}